+++ src/multi-tenancy/contracts/data-isolation-residency-portability-profile.js
@@ -1,0 +1,45 @@
+export class DataIsolationResidencyPortabilityProfile {
+  constructor({
+    profileName,
+    dataIsolationPrinciples = [], dataScope = [], dataOwnershipScopes = [], tenantOwnershipAttributeRules = [],
+    sharedRelationalStorageControls = [], partitionedDedicatedStorageRequirements = [], objectFileStorageControls = [],
+    cacheIsolationPreventions = [], searchIndexSafeguards = [], knowledgeMemoryPreventions = [], eventsMessagingControls = [],
+    observabilityDataControls = [], encryptionKeyPatterns = [], residencyPolicyFields = [], sovereigntyConstraints = [],
+    placementEnforcementChecks = [], dataTransferRequirements = [], retentionMetadataFields = [], legalHoldControls = [],
+    erasurePropagationTargets = [], backupIsolationDesignElements = [], tenantRestoreSteps = [], portabilityPackageContents = [],
+    exportArchitectureControls = [], migrationArchitectureSteps = [], dataLineageQuestions = [], dataOperationalControls = [],
+    dataTestingAssuranceRequirements = [], dataFailClosedConditions = [], contaminationResponseActions = [], architecturalRules = [],
+    tenantOwnershipExplicitForEveryAsset = true, everyProviderEnforcesOwnership = true, isolationIncludesDerivedData = true,
+    sharedStorageRequiresTenantAwareSafeguards = true, placementAndOwnershipSeparateConcepts = true,
+    residencyAppliesBeyondPrimaryStores = true, encryptionSupplementsAuthorization = true, minimizationAppliesToCopiesAndContext = true,
+    retentionAndHoldPolicyDrivenTraceable = true, erasurePropagatesToDerivatives = true, portabilityUsesGovernedManifests = true,
+    restoreAndMigrationPreserveBoundaries = true, backupsRetainTenantAttributes = true, lineageSupportsExitAndAudit = true,
+    exportJobsHaveQuotas = true, crossTenantTransferRequiresExplicitContract = true,
+    objectUrlProvesAuthorization = false, physicalIsolationEstablishesAuthorization = false,
+    keySeparationCompensatesMissingAuthorization = false, indefiniteRetentionAllowed = false, silenceEvidenceOfDeletion = false,
+    dataSilentlyReassigned = false, deidentificationByNameRemovalAlone = false, crossTenantTransferInferredFromRelationship = false
+  }) {
+    this.profileName = profileName;
+    for (const [key, value] of Object.entries({
+      dataIsolationPrinciples, dataScope, dataOwnershipScopes, tenantOwnershipAttributeRules, sharedRelationalStorageControls,
+      partitionedDedicatedStorageRequirements, objectFileStorageControls, cacheIsolationPreventions, searchIndexSafeguards,
+      knowledgeMemoryPreventions, eventsMessagingControls, observabilityDataControls, encryptionKeyPatterns,
+      residencyPolicyFields, sovereigntyConstraints, placementEnforcementChecks, dataTransferRequirements,
+      retentionMetadataFields, legalHoldControls, erasurePropagationTargets, backupIsolationDesignElements,
+      tenantRestoreSteps, portabilityPackageContents, exportArchitectureControls, migrationArchitectureSteps,
+      dataLineageQuestions, dataOperationalControls, dataTestingAssuranceRequirements, dataFailClosedConditions,
+      contaminationResponseActions, architecturalRules
+    })) this[key] = Object.freeze([...value]);
+    for (const [key, value] of Object.entries({
+      tenantOwnershipExplicitForEveryAsset, everyProviderEnforcesOwnership, isolationIncludesDerivedData,
+      sharedStorageRequiresTenantAwareSafeguards, placementAndOwnershipSeparateConcepts, residencyAppliesBeyondPrimaryStores,
+      encryptionSupplementsAuthorization, minimizationAppliesToCopiesAndContext, retentionAndHoldPolicyDrivenTraceable,
+      erasurePropagatesToDerivatives, portabilityUsesGovernedManifests, restoreAndMigrationPreserveBoundaries,
+      backupsRetainTenantAttributes, lineageSupportsExitAndAudit, exportJobsHaveQuotas, crossTenantTransferRequiresExplicitContract,
+      objectUrlProvesAuthorization, physicalIsolationEstablishesAuthorization, keySeparationCompensatesMissingAuthorization,
+      indefiniteRetentionAllowed, silenceEvidenceOfDeletion, dataSilentlyReassigned, deidentificationByNameRemovalAlone,
+      crossTenantTransferInferredFromRelationship
+    })) this[key] = Boolean(value);
+    Object.freeze(this);
+  }
+}